Rename: event dto dto/event/event로 이동
