+++ src/main/java/com/teamddd/duckmap/dto/event/event/CreateEventRes.java
@@ -1,4 +1,4 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.event.event;
 
 import lombok.Builder;
 import lombok.Getter;
