+++ src/main/java/com/teamddd/duckmap/dto/event/event/EventRes.java
@@ -1,7 +1,12 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.event.event;
 
 import java.time.LocalDateTime;
 import java.util.List;
+
+import com.teamddd.duckmap.dto.ImageRes;
+import com.teamddd.duckmap.dto.MockArtistRes;
+import com.teamddd.duckmap.dto.MockCategoryRes;
+import com.teamddd.duckmap.dto.ReviewRes;
 
 import lombok.Builder;
 import lombok.Getter;
